LayoutTests/imported/w3c/web-platform-tests/web-animations/responsive/baselineShift-expected.txt


PASS baselineShift responsive to style changes
FAIL baselineShift responsive to inherited changes assert_equals: expected "80px" but got "20px"
